# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AthletOS API",
    "description": "A governor for athletes who already train.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/shared-workouts/{token}": {
      "get": {
        "tags": [
          "workout-definitions"
        ],
        "operationId": "preview_shared_workout",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SharedWorkout"
                }
              }
            }
          }
        }
      }
    },
    "/v1/shared-workouts/{token}/copies": {
      "post": {
        "tags": [
          "workout-definitions"
        ],
        "operationId": "copy_shared_workout",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDefinition"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/v1/workout-definitions": {
      "get": {
        "tags": [
          "workout-definitions"
        ],
        "operationId": "list_workout_definitions",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefinitionList"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      },
      "post": {
        "tags": [
          "workout-definitions"
        ],
        "operationId": "create_workout_definition",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutContent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDefinition"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/v1/workout-definitions/{id}": {
      "get": {
        "tags": [
          "workout-definitions"
        ],
        "operationId": "get_workout_definition",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDefinition"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      },
      "delete": {
        "tags": [
          "workout-definitions"
        ],
        "operationId": "archive_workout_definition",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      },
      "patch": {
        "tags": [
          "workout-definitions"
        ],
        "operationId": "update_workout_definition",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDefinition"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDefinition"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/v1/workout-definitions/{id}/copies": {
      "post": {
        "tags": [
          "workout-definitions"
        ],
        "operationId": "copy_workout_definition",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDefinition"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
...
    "/v1/workout-definitions/{id}/shares": {
      "get": {
        "tags": [
          "workout-definitions"
        ],
        "operationId": "list_workout_shares",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareList"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      },
      "post": {
        "tags": [
          "workout-definitions"
        ],
        "operationId": "create_workout_share",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedShare"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/v1/workout-definitions/{id}/shares/{share_id}": {
      "delete": {
        "tags": [
          "workout-definitions"
        ],
        "operationId": "revoke_workout_share",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "share_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatedShare": {
        "type": "object",
        "required": [
          "id",
          "revision",
          "token",
          "path"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "path": {
            "type": "string"
          },
          "revision": {
            "type": "integer",
            "format": "int32"
          },
          "token": {
            "type": "string",
            "description": "Returned once; only its SHA-256 digest is stored."
          }
        }
      },
...
      "DefinitionList": {
        "type": "object",
        "required": [
          "workouts"
        ],
        "properties": {
          "workouts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkoutDefinition"
            }
          }
        }
      },
...
      "ExerciseSummary": {
        "type": "object",
        "description": "One exercise, as a client renders it in a list.",
        "required": [
          "key",
          "label",
          "is_primary"
        ],
        "properties": {
          "is_primary": {
            "type": "boolean",
            "description": "A competition lift or a close variant, as opposed to accessory work.\n\nCarried so a picker can put the lifts an athlete is likely to track a max\nfor at the top, without a client holding its own opinion about which\nthose are."
          },
          "key": {
            "type": "string",
            "description": "Stable key. This is what the maxes document is keyed by.",
            "example": "squat"
          },
          "label": {
            "type": "string",
            "example": "Squat"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 9457 problem details. `type` is omitted until we publish stable problem\ntype URIs; consumers should branch on `status` plus `title` for now.",
        "required": [
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Human-readable explanation specific to this occurrence."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code, repeated in the body per RFC 9457.",
            "example": 404,
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Short, human-readable summary of the problem type.",
            "example": "Not Found"
          }
        }
      },
...
      "RevisionRequest": {
        "type": "object",
        "required": [
          "revision"
        ],
        "properties": {
          "revision": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ShareList": {
        "type": "object",
        "required": [
          "shares"
        ],
        "properties": {
          "shares": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkoutShare"
            }
          }
        }
      },
      "SharedWorkout": {
        "allOf": [
          {
            "$ref": "#/components/schemas/WorkoutContent"
          },
          {
            "type": "object",
            "required": [
              "revision",
              "exercises"
            ],
            "properties": {
              "exercises": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ExerciseSummary"
                },
                "description": "Display metadata for this revision, also available before sign-in."
              },
              "revision": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ]
      },
...
      "UpdateDefinition": {
        "allOf": [
          {
            "$ref": "#/components/schemas/WorkoutContent"
          },
          {
            "type": "object",
            "required": [
              "expected_revision"
            ],
            "properties": {
              "expected_revision": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ]
      },
...
      "WorkoutContent": {
        "type": "object",
        "required": [
          "title",
          "blocks"
        ],
        "properties": {
          "blocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkoutBlock"
            }
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": "string"
          }
        }
      },
      "WorkoutDefinition": {
        "allOf": [
          {
            "$ref": "#/components/schemas/WorkoutContent"
          },
          {
            "type": "object",
            "required": [
              "id",
              "revision",
              "archived"
            ],
            "properties": {
              "archived": {
                "type": "boolean"
              },
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "revision": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ]
      },
...
    },
    "securitySchemes": {
      "bearer_token": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Ed25519 (EdDSA) access token from /v1/auth/login, verifiable against /.well-known/jwks.json"
      }
    }
  },
...
}